requirements.txt
# ── Core Data Engineering ──────────────────────────────────────────────────
pyspark==3.5.0
dbt-core==1.7.4
dbt-duckdb==1.7.1
apache-airflow==2.8.1
great-expectations==0.18.8

# ── Streaming ──────────────────────────────────────────────────────────────
kafka-python==2.0.2
confluent-kafka==2.3.0

# ── Storage ────────────────────────────────────────────────────────────────
boto3==1.34.0          # MinIO S3-compatible client
minio==7.2.3

# ── Data Processing ────────────────────────────────────────────────────────
pandas==2.1.4
pyarrow==14.0.2
duckdb==0.9.2

# ── Data Generation (Simulators) ──────────────────────────────────────────
faker==22.2.0

# ── Utilities ──────────────────────────────────────────────────────────────
python-dotenv==1.0.0
pydantic==2.5.3
loguru==0.7.2
tenacity==8.2.3        # Retry logic
click==8.1.7           # CLI for scripts
